Calling the function to create wordclouds and removing some duplicate code
